refactor: renamed kin.h -> error-codes.h
