+++ src/errors.c
@@ -10,7 +10,7 @@
 
 
 #include "errors.h"
-#include "kin.h"
+#include "error-codes.h"
